% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_annotation.R
\name{annotation}
\alias{annotation}
\title{Constructor for freesurfer_annotation-class}
\usage{
annotation(vertices, labels, colortable = NULL)
}
\arguments{
\item{vertices}{vector of vertices}

\item{labels}{vector of labels}

\item{colortable}{color table}
}
\description{
freesurfer_annotation is a special object
containing the three components of a 
FreeSurfer annotation file: vertices, labels
and the colortable
}
\examples{
if (have_fs()) {
 bert_dir = file.path(fs_subj_dir(), "bert")
 annot_file = file.path(bert_dir, "label", "lh.aparc.annot")
 res = read_annotation(annot_file)
    
 annotation(
   vertices = res$vertices,
   labels = res$labels,
   colortable = res$colortable
 )
} 
}
