% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats2table.R
\name{aparcstats2table.help}
\alias{aparcstats2table.help}
\title{Get Help for \code{aparcstats2table} Freesurfer Command}
\usage{
aparcstats2table.help()
}
\value{
This function does not return a value. It prints help documentation to the
console.
}
\description{
This function displays the help documentation for the Freesurfer command
\code{aparcstats2table}. The \code{aparcstats2table} command is used to extract
cortical parcellation statistics and output them in a tabular format, making
it easier to perform analyses or further processing.
}
\details{
The \code{aparcstats2table} command is commonly used when working with data
generated from Freesurfer's cortical parcellations. It allows users to
extract statistics from cortical surfaces and output them in a structured
format that can be easily analyzed or visualized.
}
\examples{
\dontshow{if (have_fs()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
aparcstats2table.help()
\dontshow{\}) # examplesIf}
}
