% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aparcstats2table.R
\name{aparcstats2table}
\alias{aparcstats2table}
\title{Parcellation Stats to Table}
\usage{
aparcstats2table(subjects, outfile = NULL, hemi = c("lh", "rh"),
  measure = c("area", "volume", "thickness", "thicknessstd", "meancurv",
  "gauscurv", "foldind", "curvind"), sep = c("tab", "space", "comma",
  "semicolon"), parc = c("aparc", "aparc.a2009s"), subj_dir = NULL,
  skip = FALSE, opts = "", verbose = TRUE)
}
\arguments{
\item{subjects}{(character) vector of subjects}

\item{outfile}{(character) output filename}

\item{opts}{(character) additional options to \code{aparcstats2table}}
}
\value{
Result of \code{system} command
}
\description{
This function calls \code{aparcstats2table} to 
convert parcellation statistics to a table
}

