% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats2table.R
\name{asegstats2table.help}
\alias{asegstats2table.help}
\title{Get Help for \code{asegstats2table} Freesurfer Command}
\usage{
asegstats2table.help()
}
\value{
This function does not return a value. It prints help documentation to the
console.
}
\description{
This function displays the help documentation for the Freesurfer command
\code{asegstats2table}. The \code{asegstats2table} command is used to extract
subcortical segmentation statistics and output them in a tabular format.
}
\details{
The \code{asegstats2table} command facilitates the analysis of data generated
from Freesurfer's automated anatomical segmentation into table formats.
}
\examples{
\dontshow{if (have_fs()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
asegstats2table.help()
\dontshow{\}) # examplesIf}
}
