% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_fs_result.R
\name{check_fs_result}
\alias{check_fs_result}
\title{Check Freesurfer Result}
\usage{
check_fs_result(res, fe_before, fe_after)
}
\arguments{
\item{res}{(numeric) Result from system command}

\item{fe_before}{(logical) did the output file exist before the command ran}

\item{fe_after}{(logical) did the output file exist after the command ran}
}
\description{
Checks the Freesurfer system command result and will
stop or warning based on whether output files exist.
}
