% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freesurfer_read3.R
\name{freesurfer_read3}
\alias{freesurfer_read3}
\title{Freesurfer Read 3 records}
\usage{
freesurfer_read3(file)
}
\arguments{
\item{file}{thickness file or anything in surf/ directory 
from Freesurfer
subject}
}
\value{
Numeric
}
\description{
Reads first 3 records of file and returns the 
rotated value,
for checking for other functions.
}
\examples{
if (have_fs()) {
   bert_dir = file.path(fs_subj_dir(), "bert", "surf")
   file = file.path(bert_dir, "lh.thickness")
   out = freesurfer_read3(file)
}
}
