% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_setup.R
\name{freesurferdir}
\alias{freesurferdir}
\alias{freesurfer_dir}
\alias{fs_dir}
\title{Get Freesurfer's Directory}
\usage{
freesurferdir()

freesurfer_dir()

fs_dir()
}
\value{
Character path to the Freesurfer home directory.
}
\description{
Finds the \code{FREESURFER_HOME} from system environment or
\code{getOption("freesurfer.home")} for location of Freesurfer functions and returns its value.
}
\examples{
\dontshow{if (have_fs()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
 freesurferdir()
 freesurfer_dir()
 fs_dir()
\dontshow{\}) # examplesIf}
}
