% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_setup.R
\name{fs_dir}
\alias{fs_dir}
\alias{freesurferdir}
\alias{freesurfer_dir}
\alias{fs_subj_dir}
\title{Get FreeSurfer Directory Paths}
\usage{
freesurferdir()

freesurfer_dir()

fs_dir()

fs_subj_dir()
}
\value{
Character; path to the FreeSurfer home or subjects directory.
}
\description{
Functions to retrieve FreeSurfer's installation and subjects directories.
Multiple aliases are provided for convenience and backward compatibility.
}
\section{Functions}{
\itemize{
\item \code{freesurferdir()}: Get FreeSurfer installation directory

\item \code{freesurfer_dir()}: Alias for freesurferdir

\item \code{fs_dir()}: Short alias for freesurferdir

\item \code{fs_subj_dir()}: Get FreeSurfer subjects directory

}}
\examples{
\dontshow{if (have_fs()) withAutoprint(\{ # examplesIf}
# All return the same value
freesurferdir()
freesurfer_dir()
fs_dir()

# Get subjects directory
fs_subj_dir()
\dontshow{\}) # examplesIf}
}
\seealso{
\code{\link[=get_fs_home]{get_fs_home()}} for more detailed information,
\code{\link[=have_fs]{have_fs()}} to check if FreeSurfer is accessible
}
