% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_help.R
\name{fs_help_rd}
\alias{fs_help_rd}
\title{Render FreeSurfer CLI help as Rd markup}
\usage{
fs_help_rd(func_name, help_arg = "--help", bin_app = "bin")
}
\arguments{
\item{func_name}{FreeSurfer function name}

\item{help_arg}{Argument to print help}

\item{bin_app}{FreeSurfer bin directory appendix}
}
\value{
Character string of Rd markup
}
\description{
Called at render time by \verb{\\Sexpr} in generated .Rd files.
Returns valid Rd markup showing CLI help or a fallback message.
}
