% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_lut.R
\docType{data}
\name{fs_lut}
\alias{fs_lut}
\title{Freesurfer look up table (LUT)}
\format{
An object of class \code{data.frame} with 1266 rows and 6 columns.
}
\usage{
fs_lut
}
\description{
A \code{data.frame} with the index, label,
and RGBA (red, blue, green, alpha) specification for the segmentations
}
\keyword{datasets}
