% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_version.R
\name{fs_version}
\alias{fs_version}
\title{Find Freesurfer Version}
\usage{
fs_version()
}
\value{
If the version file does not exist, it will throw a warning, but 
it will return an empty string.  Otherwise it will be a string of the version.
}
\description{
Finds the Freesurfer version from \code{FREESURFER_HOME/build-stamp.txt}
}
\note{
This will use \code{fs_dir()} to get the directory of FREESURFER
}
\examples{
if (have_fs()) {
 fs_version()
}
}
