% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_setup.R
\name{get_fs_output}
\alias{get_fs_output}
\title{Determine Freesurfer output type}
\usage{
get_fs_output()
}
\value{
FSF_OUTPUT_FORMAT, such as nii.gz  If none found, uses nii.gz as default
}
\description{
Finds the FSF_OUTPUT_FORMAT from system environment or 
\code{getOption("fs.outputtype")} for output type (nii.gz, nii, ANALYZE,etc)
}
\examples{
get_fs_output()
}

