% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_setup.R
\name{have_fs}
\alias{have_fs}
\title{Logical check if Freesurfer is accessible}
\usage{
have_fs(check_license = TRUE)
}
\arguments{
\item{check_license}{\link{logical} Should a license file be checked for existence?}
}
\value{
Logical \code{TRUE} if Freesurfer is accessible and license (if checked) is found, \code{FALSE} otherwise.
}
\description{
Checks if FreesurferDIR is accessible and optionally if a license file exists.
}
\examples{
have_fs()
}
