% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_annotation.R
\name{is.annotation}
\alias{is.annotation}
\alias{is_annotation}
\title{freesurfer_annotation validation}
\usage{
is.annotation(x)

is_annotation(x)
}
\arguments{
\item{x}{object to check}
}
\description{
check if object is of class freesurfer_annotation
}
\examples{
if (have_fs()) {
 bert_dir = file.path(fs_subj_dir(), "bert")
 annot_file = file.path(bert_dir, "label", "lh.aparc.annot")
 res = read_annotation(annot_file)
    
 is_annotation(res)
 is.annotation(res)
 
 is_annotation(annot_file)
 is.annotation(annot_file)
} 
}
