% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnc.R
\name{mnc2nii}
\alias{mnc2nii}
\alias{mnc2nii.help}
\title{Convert MNC to NIfTI}
\usage{
mnc2nii(file, outfile = NULL, ...)

mnc2nii.help()
}
\arguments{
\item{file}{(character) input filename}

\item{outfile}{(character) output filename}

\item{...}{Additional arguments to pass to \code{\link{fs_cmd}}}
}
\value{
Character filename of output
}
\description{
This function calls \code{mnc2nii}
to convert MNC files to NIfTI
}
\section{Functions}{
\itemize{
\item \code{mnc2nii.help()}: Display information about mnc2nii command

}}
\examples{
\dontshow{if (have_fs()) withAutoprint(\{ # examplesIf}
img = oro.nifti::nifti(array(rnorm(5*5*5), dim = c(5,5,5)))
mnc = nii2mnc(img)
img_file = mnc2nii(mnc, outfile = temp_file(fileext = ".nii"))
neurobase::readnii(img_file, verbose = get_fs_verbosity())
\dontshow{\}) # examplesIf}
}
