% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mnc.R
\name{mnc2nii.help}
\alias{mnc2nii.help}
\alias{nii2mnc.help}
\title{MNC to NIfTI Help}
\usage{
mnc2nii.help(display = TRUE, warn = TRUE, ...)

nii2mnc.help(display = TRUE, warn = TRUE, ...)
}
\arguments{
\item{display}{Logical; whether to display help output}

\item{warn}{Logical; whether to warn if help is not available}

\item{...}{Additional arguments passed to \code{\link[=fs_help]{fs_help()}}}
}
\value{
Result of \code{fs_help}
}
\description{
This calls Freesurfer's \code{mnc2nii} help
}
\section{Functions}{
\itemize{
\item \code{mnc2nii.help()}: Display information about mnc2nii command

\item \code{nii2mnc.help()}: Display information about nii2mnc command

}}
