% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mri_convert.R, R/mris_convert.R
\name{mri_convert.help}
\alias{mri_convert.help}
\title{MRI Normalize Help}
\usage{
mri_convert.help()

mri_convert.help()
}
\value{
Result of \code{fs_help}

Result of \code{fs_help}
}
\description{
This calls Freesurfer's \code{mri_convert} help

This calls Freesurfer's \code{mris_convert} help
}

