% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mri_deface.R
\name{mri_deface}
\alias{mri_deface}
\title{MRI Deface}
\usage{
mri_deface(file, brain_template = NULL, face_template = NULL, ...)
}
\arguments{
\item{file}{File to pass to \code{mri_deface}}

\item{brain_template}{\code{gca} brain template file to pass to \code{mri_deface}}

\item{face_template}{\code{gca} face template file to pass to \code{mri_deface}}

\item{...}{Additional arguments to pass to \code{\link{fs_cmd}}}
}
\value{
Result of \code{fs_cmd}, which type depends on 
arguments to \code{...}
}
\description{
This calls Freesurfer's \code{mri_deface}
}
\note{
If \code{brain_template} or\code{face_template} is \code{NULL},
they will be downloaded.
}
\examples{
if (have_fs()){
   img = oro.nifti::nifti(array(rnorm(5*5*5), dim = c(5,5,5)))  
   mri_deface(img)
}
}
