% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mri_info.R
\name{mri_info}
\alias{mri_info}
\title{MRI information}
\usage{
mri_info(file, ...)
}
\arguments{
\item{file}{File to pass to \code{mri_info}}

\item{...}{Additional arguments to pass to \code{\link{fs_cmd}}}
}
\value{
Result of \code{fs_cmd}, which type depends on 
arguments to \code{...}
}
\description{
This calls Freesurfer's \code{mri_info}
}
\examples{
if (have_fs()){
   img = oro.nifti::nifti(array(rnorm(5*5*5), dim = c(5,5,5)))  
   mri_info(img)
}
}

