% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mri_mask.R
\name{mri_mask.help}
\alias{mri_mask.help}
\title{MRI Normalize Help}
\usage{
mri_mask.help(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[=fs_help]{fs_help()}}}
}
\description{
This calls Freesurfer's \code{mri_mask} help
}
