% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mri_normalize.R
\name{mri_normalize.help}
\alias{mri_normalize.help}
\title{MRI Normalize Help}
\usage{
mri_normalize.help()
}
\value{
Result of \code{fs_help}
}
\description{
This calls Freesurfer's \code{mri_normalize} help
}

