% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mri_normalize.R
\name{mri_normalize}
\alias{mri_normalize}
\title{Use Freesurfers MRI Normalize Algorithm}
\usage{
mri_normalize(file, outfile = NULL, retimg = TRUE, opts = "", ...)
}
\arguments{
\item{file}{(character) input filename}

\item{outfile}{(character) output filename}

\item{retimg}{(logical) return image of class nifti}

\item{opts}{(character) additional options to \code{mri_normalize}}

\item{...}{additional arguments passed to \code{\link{fs_cmd}}.}
}
\value{
Character or nifti depending on \code{retimg}
}
\description{
This function calls \code{mri_normalize} to normalize the 
values of the image, with white matter voxels around 110.
}
\examples{
\dontrun{
if (have_fs()){
    mri_normalize("/path/to/T1.nii.gz")
} 
}
}
