% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mri_surf2surf.R
\name{mri_surf2surf.help}
\alias{mri_surf2surf.help}
\title{Freesurfers mri_surf2surf Help}
\usage{
mri_surf2surf.help(display = TRUE, warn = TRUE, ...)
}
\arguments{
\item{display}{Logical; whether to display help output}

\item{warn}{Logical; whether to warn if help is not available}

\item{...}{Additional arguments to pass to \code{fs_help}}
}
\value{
Result of \code{fs_help}
}
\description{
This calls Freesurfer's \code{mri_surf2surf} help
}
