% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mri_synthstrip.R
\name{mri_synthstrip.help}
\alias{mri_synthstrip.help}
\title{MRI Normalize Help}
\usage{
mri_synthstrip.help(display = TRUE, warn = TRUE, ...)
}
\arguments{
\item{display}{Logical; whether to display help output}

\item{warn}{Logical; whether to warn if help is not available}

\item{...}{additional arguments to \code{\link{get_fs}}}
}
\value{
Result of \code{fs_help}
}
\description{
This calls Freesurfer's \code{mri_mask} help
}
