% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mri_watershed.R
\name{mri_watershed.help}
\alias{mri_watershed.help}
\title{MRI Watershed Help}
\usage{
mri_watershed.help()
}
\value{
Result of \code{fs_help}
}
\description{
This calls Freesurfer's \code{mri_watershed} help
}

