% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mri_watershed.R
\name{mri_watershed}
\alias{mri_watershed}
\alias{mri_watershed.help}
\title{Use Freesurfers MRI Watershed Algorithm}
\usage{
mri_watershed(file, outfile = NULL, retimg = TRUE, opts = "", ...)

mri_watershed.help(display = TRUE, warn = TRUE, ...)
}
\arguments{
\item{file}{(character) input filename}

\item{outfile}{(character) output filename}

\item{retimg}{(logical) return image of class nifti}

\item{opts}{Character. Additional options to Freesurfer function.}

\item{...}{Additional arguments passed to \code{\link[=fs_help]{fs_help()}}}

\item{display}{Logical; whether to display help output}

\item{warn}{Logical; whether to warn if help is not available}
}
\value{
Character or nifti depending on \code{retimg}

Result of \code{fs_help}
}
\description{
This function calls \code{mri_watershed} to extract a brain
from an image, usually for skull stripping.

This calls Freesurfer's \code{mri_watershed} help
}
\section{Functions}{
\itemize{
\item \code{mri_watershed.help()}: Display FreeSurfer help for mri_watershed

}}
\examples{
\dontrun{
mri_watershed("/path/to/T1.nii.gz")
}
}
