% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mris_convert.R
\name{mris_convert.help}
\alias{mris_convert.help}
\title{Help file for Freesurfers MRIs Converter}
\usage{
mris_convert.help(display = TRUE, warn = TRUE, ...)
}
\arguments{
\item{display}{Logical; whether to display help output}

\item{warn}{Logical; whether to warn if help is not available}

\item{...}{Additional arguments passed to \code{\link[=fs_help]{fs_help()}}}
}
\description{
This calls Freesurfer's \code{mris_convert} help
}
