% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mris_convert_normals.R
\name{mris_convert_normals}
\alias{mris_convert_normals}
\title{Convert Surface to Surface normals}
\usage{
mris_convert_normals(opts = "", ...)
}
\arguments{
\item{opts}{(character) additional options to 
\code{\link{mris_convert}}}

\item{...}{additional arguments to \code{\link{mris_convert}}}
}
\value{
Result of \code{\link{mris_convert}}
}
\description{
This function call \code{mris_convert}, 
using the \code{-n} option
}
\examples{
if (have_fs()) {
 bert_dir = file.path(fs_subj_dir(), "bert")
 asc_file = mris_convert_normals(
 infile = file.path(bert_dir, "surf", "lh.white")
 )  
readLines(asc_file, n = 6)
}   
}
