% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mris_convert_vertex.R
\name{mris_convert_vertex}
\alias{mris_convert_vertex}
\title{Convert Surface to vertex file}
\usage{
mris_convert_vertex(opts = "", ...)
}
\arguments{
\item{opts}{(character) additional options to 
\code{\link{mris_convert}}}

\item{...}{additional arguments to \code{\link{mris_convert}}}
}
\value{
Result of \code{\link{mris_convert}}
}
\description{
This function call \code{mris_convert}, 
using the \code{-v} option
}
\examples{
if (have_fs()) {
 bert_surf_dir = file.path(fs_subj_dir(), "bert", "surf")
 asc_file = mris_convert_vertex(
 infile = file.path(bert_surf_dir, "lh.white")
 )  
 readLines(asc_file, n = 6)
 
} 
}
