% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mris_euler_number.R
\name{mris_euler_number.help}
\alias{mris_euler_number.help}
\title{MRI Euler Number Help}
\usage{
mris_euler_number.help(...)
}
\arguments{
\item{...}{Additional arguments passed to \code{\link[=fs_help]{fs_help()}}}
}
\description{
This calls FreeSurfer's \code{mris_euler_number} help
}
