% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nu_correct.R
\name{nu_correct.help}
\alias{nu_correct.help}
\title{Non-Uniformity Correction Help}
\usage{
nu_correct.help()
}
\value{
Result of \code{fs_help}
}
\description{
This calls Freesurfer's \code{nu_correct} help
}
