% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aparcstats2table.R
\name{read_aparcstats2table}
\alias{read_aparcstats2table}
\title{Read Parcellation Stats to Table Output}
\usage{
read_aparcstats2table(file, sep = NULL, stringsAsFactors = FALSE, ...)
}
\arguments{
\item{file}{(character path) filename of text file}

\item{sep}{separator to override attribute of file, to 
pass to \code{\link{read.table}}.}

\item{stringsAsFactors}{(logical) passed to \code{\link{read.table}}}

\item{...}{additional arguments to \code{\link{read.table}}}
}
\value{
\code{data.frame} from the file
}
\description{
This function reads a \code{aparcstats2table} output
file
}
\examples{
if (have_fs()) {
   outfile = aparcstats2table(subjects = "bert",
                    hemi = "lh",
                    meas = "thickness")
   df = read_aparcstats2table(outfile)
}
}

