% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_fs_table.R
\name{read_fs_table}
\alias{read_fs_table}
\alias{read_stats_table}
\title{Read FreeSurfer Table Output}
\usage{
read_fs_table(
  file,
  sep = NULL,
  header = TRUE,
  validate_format = TRUE,
  stringsAsFactors = FALSE,
  ...
)

read_stats_table(
  file,
  sep = NULL,
  header = TRUE,
  validate_format = TRUE,
  stringsAsFactors = FALSE,
  ...
)
}
\arguments{
\item{file}{(character path) filename of text file}

\item{sep}{separator to override attribute of file, to
pass to \code{\link[utils]{read.table}}.}

\item{header}{Is there a header in the data}

\item{validate_format}{(logical) Whether to warn about unexpected formats}

\item{stringsAsFactors}{(logical) passed to \code{\link{read.table}}}

\item{...}{additional arguments to \code{\link{read.table}}}
}
\value{
\code{data.frame} from the file
}
\description{
This function reads output from a FreeSurfer table command,
e.g. \code{aparcstats2table}, \code{asegstats2table}
}
\examples{
\dontshow{if (have_fs()) withAutoprint(\{ # examplesIf}
outfile = aparcstats2table(
   subjects = "bert",
   hemi = "lh",
   meas = "thickness"
)
df = read_fs_table(outfile)
seg_outfile = asegstats2table(subjects = "bert", meas = "mean")
df_seg = read_fs_table(seg_outfile)
\dontshow{\}) # examplesIf}
}
