% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mnc.R
\name{read_mnc}
\alias{read_mnc}
\title{Read MNC File}
\usage{
read_mnc(file)
}
\arguments{
\item{file}{(character) input filename}
}
\value{
Object of class \code{nifti}
}
\description{
This function calls \code{\link{mnc2nii}}
to convert MNC files to NIfTI, then reads it in using
\code{\link[neurobase]{readnii}}
}
