% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readmgz.R
\name{readmgz}
\alias{readmgz}
\alias{readmgh}
\title{Read MGH or MGZ File}
\usage{
readmgz(file)

readmgh(file)
}
\arguments{
\item{file}{(character) input filename}
}
\value{
Object of class \code{nifti}
}
\description{
This function calls \code{mri_convert} 
to convert MGH/MGZ files to NIfTI, then reads it in using 
\code{\link{readnii}}
}
