% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readmnc.R
\name{readmnc}
\alias{readmnc}
\title{Read MNC File}
\usage{
readmnc(file)
}
\arguments{
\item{file}{(character) input filename}
}
\value{
Object of class \code{nifti}
}
\description{
This function calls \code{mnc2nii} 
to convert MNC files to NIfTI, then reads it in using 
\code{\link{readnii}}
}
