% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_check_fs_cmd.R
\name{run_check_fs_cmd}
\alias{run_check_fs_cmd}
\title{Run and Check a Freesurfer Command}
\usage{
run_check_fs_cmd(cmd, outfile, verbose = TRUE)
}
\arguments{
\item{cmd}{Command to be run}

\item{outfile}{Output file to be produced}

\item{verbose}{print diagnostic messages}
}
\value{
Invisible NULL
}
\description{
Checks whether an output filename exists before a command
has run, prints and runs the command, and then checks the output from the
result.
}
\seealso{
\code{\link{check_fs_result}}
}
