% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_setup.R
\name{set_fs_subj_dir}
\alias{set_fs_subj_dir}
\title{Set FreeSurfer Subjects Directory}
\usage{
set_fs_subj_dir(path)
}
\arguments{
\item{path}{Character path to the subjects directory}
}
\value{
Invisibly returns the previous value of SUBJECTS_DIR
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is deprecated. Set the subjects directory using
\code{Sys.setenv(SUBJECTS_DIR = path)} or the R option
\code{options(freesurfer.subj_dir = path)} instead.
}
\keyword{internal}
