% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats2table.R
\name{stats2table}
\alias{stats2table}
\alias{asegstats2table}
\alias{aparcstats2table}
\alias{aparcstats2table.help}
\alias{asegstats2table.help}
\title{Generalized Stats to Table}
\usage{
stats2table(
  type = c("aseg", "aparc"),
  input,
  measure,
  input_type = c("subjects", "inputs"),
  outfile = NULL,
  delim = c("tab", "space", "comma", "semicolon"),
  skip = TRUE,
  subj_dir = NULL,
  opts = "",
  verbose = get_fs_verbosity(),
  ...
)

asegstats2table(
  subjects = NULL,
  inputs = NULL,
  measure = c("volume", "mean", "std"),
  ...
)

aparcstats2table(
  subjects,
  hemi = c("lh", "rh"),
  measure = c("area", "volume", "thickness", "thicknessstd", "meancurv", "gauscurv",
    "foldind", "curvind"),
  parc = c("aparc", "aparc.a2009s"),
  opts = "",
  ...
)

aparcstats2table.help(display = TRUE, warn = TRUE, ...)

asegstats2table.help(display = TRUE, warn = TRUE, ...)
}
\arguments{
\item{type}{(character) Either "aparc" for cortical parcellation or
"aseg" for subcortical segmentation.}

\item{input}{(character) A vector representing either subject IDs
(for \code{input_type = "subjects"}) or file paths
(for \code{input_type = "inputs"}, applicable only to \code{aseg}).}

\item{measure}{(character) The measurement to calculate. For example,
"thickness" for cortical measures or "volume" for subcortical.}

\item{input_type}{(character) Specifies the type of \code{input}. Must be
one of "subjects" or "inputs".}

\item{outfile}{(character) Name of the output file. If not specified,
a temporary file will be created based on the specified delimiter.}

\item{delim}{(character) Delimiter for the output file. This can be one
of: "tab", "space", "comma", or "semicolon". The output file's
delimiter is stored as an attribute for programmatic access.}

\item{skip}{(logical) If \code{TRUE}, skips invalid inputs (e.g., missing
files or data) without throwing errors.}

\item{subj_dir}{(character path) if a different subjects directory
is to be used other than \code{SUBJECTS_DIR} from shell, it can be
specified here.  Use with care as if the command fail, it may not reset
the \code{SUBJECTS_DIR} back correctly after the error}

\item{opts}{Character. Additional options to Freesurfer function.}

\item{verbose}{(logical) print diagnostic messages}

\item{...}{Additional arguments passed to \code{\link[=fs_help]{fs_help()}}}

\item{subjects}{(character) A vector of subject identifiers. This is
the primary way to specify inputs when using this function.}

\item{inputs}{(character paths) A vector of input filenames,
e.g. \code{aseg.stats}. Alternatively, use \code{subjects} to specify subject IDs.}

\item{hemi}{(character) The hemisphere for which statistics are computed.
Options are "lh" (left hemisphere) or "rh" (right hemisphere).}

\item{parc}{(character) Specifies the parcellation scheme to be used.
Options include "aparc" or "aparc.a2009s".}

\item{display}{Logical; whether to display help output}

\item{warn}{Logical; whether to warn if help is not available}
}
\value{
A character string: the path to the output file with its
delimiter stored as an attribute.
}
\description{
This function serves as a flexible abstraction to execute the FreeSurfer
commands \code{asegstats2table} or \code{aparcstats2table}, which are primarily
used to convert parcellation and segmentation statistics to tabular
formats. The function dynamically handles shared logic, constructs
command-line arguments, and runs the appropriate FreeSurfer command.
Users can specify the type of input (subjects or input file paths)
and various configuration options. With the appropriate parameters,
it constructs and executes the matching system command.
}
\section{Functions}{
\itemize{
\item \code{asegstats2table()}: Converts subcortical segmentation statistics
into tabular format by calling the FreeSurfer \code{asegstats2table}
command.
\code{...} is passed to \link{stats2table} for additional options.

\item \code{aparcstats2table()}: Converts cortical parcellation statistics
into tabular format by calling the FreeSurfer \code{aparcstats2table}
command.
\code{...} is passed to \link{stats2table} for additional options.`

\item \code{aparcstats2table.help()}: Display FreeSurfer help for aparcstats2table

\item \code{asegstats2table.help()}: Display FreeSurfer help for asegstats2table

}}
\examples{
\dontshow{if (have_fs()) withAutoprint(\{ # examplesIf}
# Example for asegstats2table
outfile_aseg <- asegstats2table(
    subjects = "bert",
    measure = "mean",
    delim = "tab"
)
print(outfile_aseg)

# Example for aparcstats2table
outfile_aparc <- aparcstats2table(
  subjects = "bert",
  hemi = "lh",
  measure = "thickness",
  delim = "tab",
  opts = "--etiv --scale=1.0"
)
print(outfile_aparc)
\dontshow{\}) # examplesIf}
}
