% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surf_convert.R
\name{surf_convert}
\alias{surf_convert}
\title{Convert Surface Data to ASCII}
\usage{
surf_convert(file, outfile = NULL, ...)
}
\arguments{
\item{file}{(character) input filename of curvature measure}

\item{outfile}{(character) output filename (if wanted to be saved)}

\item{...}{Additional arguments to pass to \code{\link{fs_cmd}}}
}
\value{
\code{data.frame}
}
\description{
This function calls \code{mri_convert} to convert a measure
from surfaces to an ASCII file and reads it in.
}
\examples{
if (have_fs()) {
   fname = file.path(fs_subj_dir(), "bert", "surf", "lh.thickness")
   out = surf_convert(fname)
}
}
