% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{temp_file}
\alias{temp_file}
\title{Create a Temporary File with a Newly Created Directory}
\usage{
temp_file(tmpdir = tempdir(check = TRUE), ...)
}
\arguments{
\item{tmpdir}{Directory in which to create the temporary file}

\item{...}{Arguments passed to \code{\link[=tempfile]{tempfile()}}}
}
\value{
Full file path to temporary file with created directory
}
\description{
Create a Temporary File with a Newly Created Directory
}
