% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trac_all.R
\name{trac_all}
\alias{trac_all}
\title{Tract Reconstruction Helper for trac-all from Freesurfer for All Steps}
\usage{
trac_all(
  infile,
  outdir = NULL,
  subjid,
  verbose = get_fs_verbosity(),
  opts = ""
)
}
\arguments{
\item{infile}{Input filename (dcm or nii)}

\item{outdir}{Character, Path to output directory.}

\item{subjid}{subject id, if NULL, the basename of the infile will be used}

\item{verbose}{(logical) print diagnostic messages}

\item{opts}{Character. Additional options to Freesurfer function.}
}
\value{
Result of \code{\link{system}}
}
\description{
Wrapper for the \code{trac-all} function in Freesurfer
for All Steps
}
