% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trac_manual.R
\name{trac_prep}
\alias{trac_prep}
\alias{trac_bedpost}
\alias{trac_path}
\title{Tract Reconstruction for Each Step}
\usage{
trac_prep(infile, outdir = NULL, subjid, verbose = get_fs_verbosity())

trac_bedpost(infile, outdir = NULL, subjid, verbose = get_fs_verbosity())

trac_path(infile, outdir = NULL, subjid, verbose = get_fs_verbosity())
}
\arguments{
\item{infile}{Input filename (dcm or nii)}

\item{outdir}{Character, Path to output directory.}

\item{subjid}{subject id, if NULL, the basename of the infile will be used}

\item{verbose}{(logical) print diagnostic messages}
}
\value{
Result of \code{\link{system}}
}
\description{
Reconstruction from Freesurfer for Preprocessing,
Bedpost, and Path reconstruction
}
