% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aparcstats2table.R
\name{aparcstats2table.help}
\alias{aparcstats2table.help}
\title{Parcellation Stats to Table Help}
\usage{
aparcstats2table.help()
}
\value{
Result of \code{fs_help}
}
\description{
This calls Freesurfer's \code{aparcstats2table} help
}
