% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asegstats2table.R
\name{asegstats2table.help}
\alias{asegstats2table.help}
\title{Parcellation Stats to Table Help}
\usage{
asegstats2table.help()
}
\value{
Result of \code{fs_help}
}
\description{
This calls Freesurfer's \code{asegstats2table} help
}

