% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkmnc.R
\docType{methods}
\name{checkmnc-methods}
\alias{checkmnc-methods}
\alias{checkmnc}
\alias{checkmnc,nifti-method}
\alias{checkmnc,character-method}
\alias{checkmnc,list-method}
\alias{ensure_mnc}
\title{Force object to filename with .mnc extension}
\usage{
checkmnc(file, ...)

\S4method{checkmnc}{nifti}(file, ...)

\S4method{checkmnc}{character}(file, ...)

\S4method{checkmnc}{list}(file, ...)

ensure_mnc(file, ...)
}
\arguments{
\item{file}{character or \code{nifti} object}

\item{...}{options passed to \code{\link[neurobase]{checkimg}}}
}
\value{
Character filename of mnc image
}
\description{
Ensures the output to be a character filename (or vector) 
from an input image or \code{nifti} to have \code{.mnc} extension and
be converted to MNC when necessary
}
\author{
John Muschelli \email{muschellij2@gmail.com}
}
