% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freesurfer-package.R
\docType{package}
\name{freesurfer-package}
\alias{freesurfer}
\alias{freesurfer-package}
\title{freesurfer: Wrapper Functions for 'FreeSurfer'}
\description{
Wrapper functions that interface with 'Freesurfer' \url{https://surfer.nmr.mgh.harvard.edu/}, a powerful and commonly-used 'neuroimaging' software, using system commands. The goal is to be able to interface with 'Freesurfer' completely in R, where you pass R objects of class 'nifti', implemented by package 'oro.nifti', and the function executes an 'Freesurfer' command and returns an R object of class 'nifti' or necessary output.
}
\seealso{
Useful links:
\itemize{
  \item Report bugs at \url{https://github.com/muschellij2/freesurfer/issues}
}

}
\author{
\strong{Maintainer}: John Muschelli \email{muschellij2@gmail.com} (\href{https://orcid.org/0000-0001-6469-1750}{ORCID}) [copyright holder]

Other contributors:
\itemize{
  \item Athanasia Mo Mowinckel (\href{https://orcid.org/0000-0002-5756-0223}{ORCID}) [contributor]
}

}
\keyword{internal}
