% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freesurfer_read3.R
\name{freesurfer_read_surf}
\alias{freesurfer_read_surf}
\title{Read Freesurfer Surface file}
\usage{
freesurfer_read_surf(file)
}
\arguments{
\item{file}{surface file (e.g. \code{lh.inflated})}
}
\value{
List of length 2: vertices and faces are the elements
}
\description{
Reads a Freesurfer Surface file from the \code{surf/} directory
from \code{recon-all}
}
\examples{
if (have_fs()) {
   fname = file.path(fs_subj_dir(), "bert", "surf", "lh.inflated")
   out = freesurfer_read_surf(fname)
}
}

