% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_setup.R
\docType{data}
\name{freesurferdir}
\alias{freesurferdir}
\alias{freesurfer_dir}
\alias{fs_dir}
\title{Get Freesurfer's Directory}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
freesurferdir()

freesurfer_dir

fs_dir
}
\value{
Character path
}
\description{
Finds the FREESURFER_HOME from system environment or
\code{getOption("freesurfer.path")}
for location of Freesurfer functions and returns it
}
\examples{
\dontshow{if (have_fs()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
freesurferdir()
freesurfer_dir()
fs_dir()
\dontshow{\}) # examplesIf}
}
\keyword{datasets}
