% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_setup.R
\name{freesurferdir}
\alias{freesurfer_dir}
\alias{freesurferdir}
\alias{fs_dir}
\title{Get Freesurfer's Directory}
\usage{
freesurferdir()

freesurfer_dir()

fs_dir()
}
\value{
Character path
}
\description{
Finds the FreesurferDIR from system environment or \code{getOption("freesurfer.path")}
for location of Freesurfer fuctions and returns it
}
\examples{
if (have_fs()) {
 freesurferdir()
 freesurfer_dir()
 fs_dir()
}
}

