% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_help.R
\name{fs_help}
\alias{fs_help}
\title{Wrapper for getting Freesurfer help}
\usage{
fs_help(func_name, help.arg = "--help", extra.args = "", ...)
}
\arguments{
\item{func_name}{Freesurfer function name}

\item{help.arg}{Argument to print help, usually "--help"}

\item{extra.args}{Extra arguments to be passed other than 
\code{--help}}

\item{...}{additional arguments to \code{\link{get_fs}}}
}
\value{
Prints help output and returns output as character vector
}
\description{
This function takes in the function and returns the
help from Freesurfer for that function
}
\examples{
if (have_fs()) {
fs_help(func_name = "mri_watershed")
}
}
