% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_setup.R
\name{fs_imgext}
\alias{fs_imgext}
\title{Determine extension of image based on FSLOUTPUTTYPE}
\usage{
fs_imgext()
}
\value{
Extension for output type
}
\description{
Runs \code{get_fs_output()} to extract FSLOUTPUTTYPE and then 
gets corresponding extension (such as .nii.gz)
}
\examples{
fs_imgext()
}

