% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_sitrep.R
\name{fs_sitrep}
\alias{fs_sitrep}
\title{Freesurfer Situation Report}
\usage{
fs_sitrep()
}
\description{
Get a situation report on your current Freesurfer installation and settings
within R. This function helps diagnose problems by showing how Freesurfer
paths and options are determined.
}
\examples{
\dontrun{
# Report all Freesurfer settings
fs_sitrep()
}
}
