% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_setup.R
\name{fs_subj_dir}
\alias{fs_subj_dir}
\title{Determine Freesurfer Subjects Directory}
\usage{
fs_subj_dir()
}
\value{
SUBJECTS_DIR, such as \code{${FREESURFER_HOME}/subjects}
}
\description{
Finds the SUBJECTS_DIR from system environment or 
\code{getOption("fs.subj_dir")} for subjects dir
}
\examples{
if (have_fs()) {
   fs_subj_dir()
}
}

