% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mri_convert.R
\name{mri_convert}
\alias{mri_convert}
\title{Use Freesurfers MRI Converter}
\usage{
mri_convert(file, outfile, opts = "")
}
\arguments{
\item{file}{(character) input filename}

\item{outfile}{(character) output filename}

\item{opts}{(character) additional options to \code{mri_convert}}

\item{...}{additional arguments passed to \code{\link{fs_cmd}}.}
}
\value{
Result of \code{system} command
}
\description{
This function calls \code{mri_convert} to convert an image
}

