% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mri_convert.R
\name{mri_convert}
\alias{mri_convert}
\title{Use Freesurfers MRI Converter}
\usage{
mri_convert(file, outfile, opts = "", ...)
}
\arguments{
\item{file}{(character) input filename}

\item{outfile}{(character) output filename}

\item{opts}{Character. Additional options to Freesurfer function.}

\item{...}{Additional arguments to pass to \code{\link{fs_cmd}}}
}
\value{
Result of \code{system} command
}
\description{
This function calls \code{mri_convert} to convert an image
}
\examples{
\dontshow{if (have_fs()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
img = oro.nifti::nifti(array(rnorm(5*5*5), dim = c(5,5,5)))
res = mri_convert(img, outfile = temp_file(fileext = ".mgz"))
\dontshow{\}) # examplesIf}
}
