% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mri_convert.R
\name{mri_convert}
\alias{mri_convert}
\title{Use Freesurfers MRI Converter}
\usage{
mri_convert(file, outfile, opts = "")
}
\arguments{
\item{file}{(character) input filename}

\item{outfile}{(character) output filename}

\item{opts}{(character) additional options to \code{mri_convert}}
}
\value{
Result of \code{system} command
}
\description{
This function calls \code{mri_convert} to convert an image
}
\examples{
if (have_fs()) {
   img = oro.nifti::nifti(array(rnorm(5*5*5), dim = c(5,5,5))) 
   res = mri_convert(img, outfile = tempfile(fileext = ".mgz"))
} 
}
