% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mri_info.R
\name{mri_info.help}
\alias{mri_info.help}
\title{MRI information Help}
\usage{
mri_info.help(display = TRUE, warn = TRUE, ...)
}
\arguments{
\item{display}{Logical; whether to display help output}

\item{warn}{Logical; whether to warn if help is not available}

\item{...}{Additional arguments passed to \code{\link[=fs_help]{fs_help()}}}
}
\description{
This calls Freesurfer's \code{mri_info} help
}
