% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mri_info.R
\name{mri_info.help}
\alias{mri_info.help}
\title{MRI information Help}
\usage{
mri_info.help()
}
\arguments{
\item{file}{File to pass to \code{mri_info}}

\item{...}{Additional arguments to pass to \code{\link{fs_cmd}}}
}
\value{
Result of \code{fs_cmd}, which type depends on 
arguments to \code{...}
}
\description{
This calls Freesurfer's \code{mri_info} help
}

