% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mri_segment.R
\name{mri_segment}
\alias{mri_segment}
\title{Use Freesurfers MRI Segmentation Algorithm}
\usage{
mri_segment(file, outfile = NULL, retimg = TRUE, opts = "", ...)
}
\arguments{
\item{file}{(character) input filename}

\item{outfile}{(character) output filename}

\item{retimg}{(logical) return image of class nifti}

\item{opts}{(character) additional options to \code{mri_segment}}

\item{...}{additional arguments passed to \code{\link{fs_cmd}}.}
}
\value{
Character or nifti depending on \code{retimg}
}
\description{
This function calls \code{mri_segment} 
to segment tissues from an image
}
\note{
NOT COMPLETE
}
\examples{
\dontrun{
if (have_fs()){
    mri_segment("/path/to/T1.nii.gz")
} 
} 
}

