% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mri_surf2surf.R
\name{mri_surf2surf.help}
\alias{mri_surf2surf.help}
\title{Freesurfers mri_surf2surf Help}
\usage{
mri_surf2surf.help()
}
\value{
Result of \code{fs_help}
}
\description{
This calls Freesurfer's \code{mri_surf2surf} help
}

