% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mri_synthstrip.R
\name{mri_synthstrip.help}
\alias{mri_synthstrip.help}
\title{MRI Normalize Help}
\usage{
mri_synthstrip.help()
}
\value{
Result of \code{fs_help}
}
\description{
This calls Freesurfer's \code{mri_mask} help
}
