% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mri_watershed.R
\name{mri_watershed}
\alias{mri_watershed}
\title{Use Freesurfers MRI Watershed Algorithm}
\usage{
mri_watershed(file, outfile = NULL, retimg = TRUE, opts = "", ...)
}
\arguments{
\item{file}{(character) input filename}

\item{outfile}{(character) output filename}

\item{retimg}{(logical) return image of class nifti}

\item{opts}{(character) additional options to \code{mri_watershed}}

\item{...}{additional arguments passed to \code{\link{fs_cmd}}.}
}
\value{
Character or nifti depending on \code{retimg}
}
\description{
This function calls \code{mri_watershed} to extract a brain 
from an image, usually for skull stripping.
}
\examples{
\dontrun{
if (have_fs()){
    mri_watershed("/path/to/T1.nii.gz")
} 
}  
}
