% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mris_convert.R
\name{mris_convert.help}
\alias{mris_convert.help}
\title{Help file for Freesurfers MRIs Converter}
\usage{
mris_convert.help()
}
\value{
Result of \code{fs_help}
}
\description{
This calls Freesurfer's \code{mris_convert} help
}

