% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mris_convert.R
\name{mris_convert}
\alias{mris_convert}
\title{Use Freesurfers MRIs Converter}
\usage{
mris_convert(curvinfilepath, origfilepath, outfile = NULL, ext = ".asc")
}
\arguments{
\item{curvinfilepath}{(character) input filename for curve}

\item{origfilepath}{(character) file path for origin file}

\item{outfile}{(character) output file path}
}
\value{
Result of \code{system} command
}
\description{
This function call  \code{mris_convert}, a general conversion program for converting between cortical surface file formats
}
\examples{
if (have_fs()) {
 mris_convert(curvinfilepath = 'lh.thickness', origfilepath = 'lh.white', outfile = 'lh.thickness.asc')  
} 
}

