% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nu_correct.R
\name{nu_correct}
\alias{nu_correct}
\title{Use Freesurfers Non-Uniformity Correction}
\usage{
nu_correct(file, outfile = NULL, opts = "", ...)
}
\arguments{
\item{file}{(character) input filename}

\item{outfile}{(character) output filename}

\item{opts}{(character) additional options to \code{mri_segment}}

\item{...}{additional arguments passed to \code{\link{fs_cmd}}.}
}
\value{
Character or nifti depending on \code{retimg}
}
\description{
This function calls \code{nu_correct} 
to correct for non-uniformity
}

