% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nu_correct.R
\name{nu_correct}
\alias{nu_correct}
\title{Use Freesurfers Non-Uniformity Correction}
\usage{
nu_correct(file, mask = NULL, opts = "", verbose = TRUE, ...)
}
\arguments{
\item{file}{(character) input filename}

\item{mask}{(character or nifti) Mask to use for correction.}

\item{opts}{(character) additional options to \code{mri_segment}}

\item{verbose}{print diagnostic messages}

\item{...}{additional arguments passed to \code{\link{fs_cmd}}.}
}
\value{
Object of class nifti depending on \code{retimg}
}
\description{
This function calls \code{nu_correct}
to correct for non-uniformity
}
\examples{
\dontrun{
if (have_fs()){
    nu_correct("/path/to/T1.nii.gz")
}
}
}
