% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aseg_stats.R
\name{read_aseg_stats}
\alias{read_aseg_stats}
\title{Read Anatomical Segmentation Statistics}
\usage{
read_aseg_stats(file, lowercase = TRUE)
}
\arguments{
\item{file}{aseg.stats file from Freesurfer}

<<<<<<< HEAD
\item{lowercase}{should the measures and values be lowercase (\code{TRUE}) or
=======
\item{lowercase}{should the measures and values be lowercase (`TRUE`) or
>>>>>>> a9fbf4b (use roxygen @examplesIf for tests)
kept as is?}
}
\value{
List of 2 \code{data.frame}s, one with the global measures and one
with the structure-specific measures.
}
\description{
Reads an \code{aseg.stats} file from an individual subject
}
\examples{
\dontshow{if (have_fs()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
 file = file.path(fs_subj_dir(), "bert", "stats", "aseg.stats")
 out = read_aseg_stats(file)
\dontshow{\}) # examplesIf}
}
