% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_aseg_stats.R
\name{read_aseg_stats}
\alias{read_aseg_stats}
\title{Read Anatomical Segmentation Statistics}
\usage{
read_aseg_stats(file, lowercase = TRUE)
}
\arguments{
\item{file}{aseg.stats file from Freesurfer}

\item{lowercase}{should the measures and values be lowercase (`TRUE`) or 
kept as is?}
}
\value{
List of 2 \code{data.frame}s, one with the global measures and one 
with the structure-specific measures.
}
\description{
Reads an \code{aseg.stats} file from an individual subject
}
\examples{
if (have_fs()) {
 file = file.path(fs_subj_dir(), "bert", "stats", "aseg.stats")
 out = read_aseg_stats(file)
}
}
