% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_mgz.R
\name{read_mgz}
\alias{read_mgz}
\alias{read_mgh}
\title{Read MGH or MGZ File}
\usage{
read_mgz(file, ...)

read_mgh(file, ...)
}
\arguments{
\item{file}{(character) input filename}

\item{...}{Additional arguments to pass to \code{\link{fs_cmd}}}
}
\value{
Object of class \code{nifti}
}
\description{
This function calls \code{mri_convert}
to convert MGH/MGZ files to NIfTI, then reads it in using
\code{\link[neurobase]{readnii}}
}
