% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recon_all.R
\name{recon_all}
\alias{recon_all}
\title{Reconstruction from Freesurfer for All Steps}
\usage{
recon_all(
  infile = NULL,
  outdir = NULL,
  subjid = NULL,
  verbose = get_fs_verbosity(),
  opts = "-all",
  ...
)
}
\arguments{
\item{infile}{Input filename (dcm or nii)}

\item{outdir}{Character, Path to output directory.}

\item{subjid}{subject id, if NULL, the basename of the infile will be used}

\item{verbose}{(logical) print diagnostic messages}

\item{opts}{Character. Additional options to Freesurfer function.}

\item{...}{arguments passed to \code{\link{reconner}}}
}
\value{
Result of \code{\link{system}}
}
\description{
Reconstruction from Freesurfer for All Steps
}
\note{
If you would like to restart a \code{recon-all} run,
change opts so that \code{opts = "-make all"}
}
