% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recon_all.R
\name{recon_all}
\alias{recon_all}
\title{Reconstruction from Freesurfer}
\usage{
recon_all(infile, outdir, subjid, verbose = TRUE, opts = "-all")
}
\arguments{
\item{infile}{Input filename (dcm or nii)}

\item{outdir}{Output directory}

\item{subjid}{subject id}

\item{verbose}{print diagnostic messages}

\item{opts}{Additional options}
}
\value{
Result of \code{\link{system}}
}
\description{
Reconstruction from Freesurfer with most of the options
implemented.
}

