% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recon_all.R
\name{recon_all}
\alias{recon_all}
\title{Reconstruction from Freesurfer for All Steps}
\usage{
recon_all(infile = NULL, outdir = NULL, subjid = NULL, verbose = TRUE,
  opts = "-all", ...)
}
\arguments{
\item{infile}{Input filename (dcm or nii)}

\item{outdir}{Output directory}

\item{subjid}{subject id}

\item{verbose}{print diagnostic messages}

\item{opts}{Additional options}

\item{...}{arguments passed to \code{\link{reconner}}}
}
\value{
Result of \code{\link{system}}
}
\description{
Reconstruction from Freesurfer for All Steps
}
\note{
If you would like to restart a \code{recon-all} run,
change opts so that \code{opts = "-make all"}
}
