% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recon_manual.R
\name{recon_con1}
\alias{autorecon1}
\alias{autorecon2}
\alias{autorecon3}
\alias{recon_con1}
\alias{recon_con1,recon_con2,recon_con3}
\alias{recon_con2}
\alias{recon_con3}
\title{Reconstruction from Motion Correction to Skull Strip}
\usage{
recon_con1(infile, outdir = NULL, subjid, verbose = TRUE)

autorecon1(infile, outdir = NULL, subjid, verbose = TRUE)

recon_con2(infile, outdir = NULL, subjid, verbose = TRUE)

autorecon2(infile, outdir = NULL, subjid, verbose = TRUE)

recon_con3(infile, outdir = NULL, subjid, verbose = TRUE)

autorecon3(infile, outdir = NULL, subjid, verbose = TRUE)
}
\arguments{
\item{infile}{Input filename (dcm or nii)}

\item{outdir}{Output directory}

\item{subjid}{subject id}

\item{verbose}{print diagnostic messages}
}
\value{
Result of \code{\link{system}}
}
\description{
Reconstruction from Freesurfer for Step 1-5 
(Motion Correction to Skull Strip), which calls \code{-autorecon1}
in \code{recon-all}
}
\note{
See https://surfer.nmr.mgh.harvard.edu/fswiki/recon-all for the 
steps of each \code{autorecon1-3}. 
If you set \code{infile = NULL}, then you can omit the 
\code{-i} flag in \code{recon-all}.
}

