% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconner.R
\name{reconner}
\alias{reconner}
\title{Reconstruction Helper for recon from Freesurfer}
\usage{
reconner(
  infile = NULL,
  outdir = NULL,
  subjid = NULL,
  verbose = get_fs_verbosity(),
  opts = "-all",
  force = FALSE
)
}
\arguments{
\item{infile}{Input filename (dcm or nii)}

\item{outdir}{Character, Path to output directory.}

\item{subjid}{subject id, if NULL, the basename of the infile will be used}

\item{verbose}{(logical) print diagnostic messages}

\item{opts}{Character. Additional options to Freesurfer function.}

\item{force}{Force running of the reconstruction}
}
\value{
Result of \code{\link{system}}
}
\description{
Wrapper for the \code{recon-all} function in Freesurfer
}
\note{
If you set \code{infile = NULL}, then you can omit the
\code{-i} flag in \code{recon-all}
}
