% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reconner.R
\name{reconner}
\alias{reconner}
\title{Reconstruction Helper for recon from Freesurfer}
\usage{
reconner(infile, outdir, subjid, verbose = TRUE, opts = "-all")
}
\arguments{
\item{infile}{Input filename (dcm or nii)}

\item{outdir}{Output directory}

\item{subjid}{subject id}

\item{verbose}{print diagnostic messages}

\item{opts}{Additional options}
}
\value{
Result of \code{\link{system}}
}
\description{
Wrapper for the \code{recon-all} function in Freesurfer
}
\note{
If you set \code{infile = NULL}, then you can omit the 
\code{-i} flag in \code{recon-all}
}

