% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_setup.R
\name{set_fs_subj_dir}
\alias{set_fs_subj_dir}
\title{Set Freesurfer Subjects Directory}
\usage{
set_fs_subj_dir(x = file.path(fs_dir(), "subjects"))
}
\arguments{
\item{x}{path to SUBJECTS_DIR defaults to \code{file.path(fs_dir(), "subjects")}}
}
\description{
Sets the SUBJECTS_DIR variable in the system environment or 
\code{options("fs.subj_dir" = x)}
}

