% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trac_all.R
\name{trac_all}
\alias{trac_all}
\title{Tract Reconstruction Helper for trac-all from Freesurfer for All Steps}
\usage{
trac_all(infile, outdir = NULL, subjid, verbose = TRUE, opts = "")
}
\arguments{
\item{infile}{Input filename (dcm or nii)}

\item{outdir}{Output directory}

\item{subjid}{subject id}

\item{verbose}{print diagnostic messages}

\item{opts}{Additional options}
}
\value{
Result of \code{\link{system}}
}
\description{
Wrapper for the \code{trac-all} function in Freesurfer 
for All Steps
}

