% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trac_manual.R
\name{trac_prep}
\alias{trac_bedpost}
\alias{trac_path}
\alias{trac_prep}
\title{Tract Reconstruction for Each Step}
\usage{
trac_prep(infile, outdir = NULL, subjid, verbose = TRUE)

trac_bedpost(infile, outdir = NULL, subjid, verbose = TRUE)

trac_path(infile, outdir = NULL, subjid, verbose = TRUE)
}
\arguments{
\item{infile}{Input filename (dcm or nii)}

\item{outdir}{Output directory}

\item{subjid}{subject id}

\item{verbose}{print diagnostic messages}
}
\value{
Result of \code{\link{system}}
}
\description{
Reconstruction from Freesurfer for Preprocessing,
Bedpost, and Path reconstruction
}

